# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TEQUALY Field Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compromissos:
    get:
      operationId: listCompromissos
      tags: [compromissos]
      summary: List commitments
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Compromisso"
    post:
      operationId: createCompromisso
      tags: [compromissos]
      summary: Create a commitment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompromissoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compromisso"
  /compromissos/{id}:
    patch:
      operationId: updateCompromisso
      tags: [compromissos]
      summary: Update commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompromissoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compromisso"
    delete:
      operationId: deleteCompromisso
      tags: [compromissos]
      summary: Delete commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Documentos ───────────────────────────────────────────
  /documentos:
    get:
      operationId: listDocumentos
      tags: [documentos]
      summary: List documents
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: tipo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Documento"
    post:
      operationId: createDocumento
      tags: [documentos]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
  /documentos/{id}:
    patch:
      operationId: updateDocumento
      tags: [documentos]
      summary: Update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
    delete:
      operationId: deleteDocumento
      tags: [documentos]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Conciliações ─────────────────────────────────────────
  /conciliacoes:
    get:
      operationId: listConciliacoes
      tags: [conciliacoes]
      summary: List reconciliations
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conciliacao"
    post:
      operationId: createConciliacao
      tags: [conciliacoes]
      summary: Create a reconciliation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConciliacaoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conciliacao"
  /conciliacoes/{id}:
    patch:
      operationId: updateConciliacao
      tags: [conciliacoes]
      summary: Update reconciliation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConciliacaoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conciliacao"
    delete:
      operationId: deleteConciliacao
      tags: [conciliacoes]
      summary: Delete reconciliation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /excecoes:
    get:
      operationId: listExcecoes
      tags: [excecoes]
      summary: List exceptions
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Excecao"
    post:
      operationId: createExcecao
      tags: [excecoes]
      summary: Create exception
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExcecaoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Excecao"
  /excecoes/{id}:
    patch:
      operationId: updateExcecao
      tags: [excecoes]
      summary: Update exception
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExcecaoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Excecao"
    delete:
      operationId: deleteExcecao
      tags: [excecoes]
      summary: Delete exception
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── KPIs ─────────────────────────────────────────────────
  /kpis:
    get:
      operationId: listKpis
      tags: [kpis]
      summary: List KPIs
      parameters:
        - name: obra_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: periodo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
    post:
      operationId: createKpi
      tags: [kpis]
      summary: Create KPI entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
  /kpis/{id}:
    patch:
      operationId: updateKpi
      tags: [kpis]
      summary: Update KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    delete:
      operationId: deleteKpi
      tags: [kpis]
      summary: Delete KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /audit-log:
    get:
      operationId: listAuditLog
      tags: [audit]
      summary: List audit log entries
      parameters:
        - name: entidade
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    Compromisso:
      type: object
      required: [id, obra_id, descricao, status, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        descricao:
          type: string
        responsavel:
          type: ["string", "null"]
        prazo:
          type: ["string", "null"]
        status:
          type: string
        prioridade:
          type: ["string", "null"]
        created_at:
          type: string
    CompromissoInput:
      type: object
      required: [obra_id, descricao, status]
      properties:
        obra_id:
          type: integer
        descricao:
          type: string
        responsavel:
          type: string
        prazo:
          type: string
        status:
          type: string
        prioridade:
          type: string
    CompromissoUpdate:
      type: object
      properties:
        descricao:
          type: string
        responsavel:
          type: string
        prazo:
          type: string
        status:
          type: string
        prioridade:
          type: string
    Documento:
      type: object
      required: [id, obra_id, nome, tipo, status, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        nome:
          type: string
        tipo:
          type: string
        url:
          type: ["string", "null"]
        versao:
          type: ["string", "null"]
        status:
          type: string
        created_at:
          type: string
    DocumentoInput:
      type: object
      required: [obra_id, nome, tipo, status]
      properties:
        obra_id:
          type: integer
        nome:
          type: string
        tipo:
          type: string
        url:
          type: string
        versao:
          type: string
        status:
          type: string
    DocumentoUpdate:
      type: object
      properties:
        nome:
          type: string
        tipo:
          type: string
        url:
          type: string
        versao:
          type: string
        status:
          type: string
    Conciliacao:
      type: object
      required: [id, obra_id, descricao, status, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        descricao:
          type: string
        valor:
          type: ["number", "null"]
        status:
          type: string
        data_referencia:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        created_at:
          type: string
    ConciliacaoInput:
      type: object
      required: [obra_id, descricao, status]
      properties:
        obra_id:
          type: integer
        descricao:
          type: string
        valor:
          type: number
        status:
          type: string
        data_referencia:
          type: string
        observacoes:
          type: string
    ConciliacaoUpdate:
      type: object
      properties:
        descricao:
          type: string
        valor:
          type: number
        status:
          type: string
        data_referencia:
          type: string
        observacoes:
          type: string
    Excecao:
      type: object
      required: [id, obra_id, descricao, status, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        conciliacao_id:
          type: ["integer", "null"]
        conciliacao_item_id:
          type: ["integer", "null"]
        descricao:
          type: string
        motivo:
          type: ["string", "null"]
        tipo_divergencia:
          type: ["string", "null"]
        fonte:
          type: ["string", "null"]
        confianca:
          type: ["string", "null"]
        acao_humana_necessaria:
          type: ["string", "null"]
        status:
          type: string
        aprovador:
          type: ["string", "null"]
        data_aprovacao:
          type: ["string", "null"]
        created_at:
          type: string
    ExcecaoInput:
      type: object
      required: [obra_id, descricao, status]
      properties:
        obra_id:
          type: integer
        descricao:
          type: string
        motivo:
          type: string
        status:
          type: string
        aprovador:
          type: string
    ExcecaoUpdate:
      type: object
      properties:
        descricao:
          type: string
        motivo:
          type: string
        status:
          type: string
        aprovador:
          type: string
        data_aprovacao:
          type: string
...
    Kpi:
      type: object
      required: [id, obra_id, nome, valor, unidade, periodo, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        nome:
          type: string
        valor:
          type: number
        unidade:
          type: string
        periodo:
          type: string
        meta:
          type: ["number", "null"]
        created_at:
          type: string
    KpiInput:
      type: object
      required: [obra_id, nome, valor, unidade, periodo]
      properties:
        obra_id:
          type: integer
        nome:
          type: string
        valor:
          type: number
        unidade:
          type: string
        periodo:
          type: string
        meta:
          type: number
    KpiUpdate:
      type: object
      properties:
        nome:
          type: string
        valor:
          type: number
        unidade:
          type: string
        periodo:
          type: string
        meta:
          type: number
...
    AuditLog:
      type: object
      required: [id, acao, entidade, created_at]
      properties:
        id:
          type: integer
        usuario_id:
          type: ["string", "null"]
        usuario_nome:
          type: ["string", "null"]
        acao:
          type: string
        entidade:
          type: string
        entidade_id:
          type: ["integer", "null"]
        detalhes:
          type: ["string", "null"]
        created_at:
          type: string