# lib/api-zod/src/generated/types/conciliacaoInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TEQUALY Field Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ConciliacaoInput {
  obra_id: number;
  descricao: string;
  valor?: number;
  status: string;
  data_referencia?: string;
  observacoes?: string;
}